requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
bidict==0.22.1
black==23.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
executing==1.1.1
Flask==2.2.2
Flask-SocketIO==5.3.5
gunicorn==21.2.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-decouple==3.8
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.28.1
six==1.16.0
stack-data==0.5.1
tomli==2.0.1
traitlets==5.5.0
typing_extensions==4.7.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.9.0
